Moved logic from NewPost.tsx to new useNewPost hook.
